### 2025-01-28 #7714 @crvernon
:wave: @Mmasoud1 - Thank you for your submission to JOSS.  I am going to run this one by our larger editorial board for review to ensure it meets our substantial scholarly effort requirements.  I'll be back in touch ASAP.
