### 2025-02-10 3da751d
Add Github Actions CI for testing
