### 2025-02-10 b060706
Add Github Actions CI for testing
