+++ .github/workflows/draft-pdf.yml
@@ -13,7 +13,7 @@ jobs:
           journal: joss
           paper-path: paper/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v3
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           path: paper/paper.pdf
