+++ .github/workflows/pytest.yml
@@ -1,32 +1,0 @@
-name: Run Tests with pytest
-
-# Trigger the workflow on pull request events
-on:
-  pull_request:
-    branches:
-      - main  
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.8, 3.9]
-
-    steps:
-      - name: Check out the repository
-        uses: actions/checkout@v2
-
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install pytest
-
-      - name: Run tests with pytest
-        run: |
-          pytest
