### 2025-02-10 5c72bdb
Add Github Actions CI for testing
