### 2025-02-10 3735d5c
Add Github Actions CI for testing
