### 2025-02-10 52c3ef0
Fix Github Actions CI for testing
