+++ requirements.txt
@@ -1,0 +1,2 @@
+pytest==8.3.4
+torch==2.0.1
