+++ .github/workflows/pytest.yml
@@ -1,6 +1,5 @@
 name: Run Tests with pytest
 
-# Trigger the workflow on pull request events
 on:
   push:
     branches: [main]
@@ -36,6 +35,11 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements.txt
 
+      - name: Install your package in editable mode
+        run: |
+          pip install -e .
+
       - name: Run tests with pytest
         run: |
           pytest
+
