+++ tests/test_dist.py
@@ -1,0 +1,168 @@
+import os
+import tempfile
+import torch
+import logging
+import numpy as np
+import pytest
+from app.code.executor.dist import training, GenericLogger
+
+# ----------------------------
+# Dummy implementations
+# ----------------------------
+
+# Dummy MeshNet: A minimal network that uses a single Conv3d.
+class DummyMeshNet(torch.nn.Module):
+    def __init__(self, in_channels, n_classes, channels, config_file):
+        super(DummyMeshNet, self).__init__()
+        # A single convolution that preserves spatial dimensions.
+        self.conv = torch.nn.Conv3d(in_channels, n_classes, kernel_size=1)
+    def forward(self, x):
+        return self.conv(x)
+# We use DummyMeshNet as our enMesh_checkpoint.
+DummyMeshNetCheckpoint = DummyMeshNet
+
+# Dummy Loader: Returns a DataLoader from a simple dataset.
+class DummyLoader:
+    class DummyDataset(torch.utils.data.Dataset):
+        def __init__(self, length=4):
+            self.length = length
+            self.device = torch.device("cpu")
+        def __len__(self):
+            return self.length
+        def __getitem__(self, idx):
+            # Use a small shape for testing purposes.
+            # The training class expects a tensor reshaped to (-1, 1, shape, shape, shape),
+            # where shape = 256 // cubes. For cubes=1, shape is 256, but we use a smaller value.
+            shape = 8
+            # Create a dummy image (with an extra channel dimension) and label.
+            image = torch.ones((1, shape, shape, shape))
+            label = torch.zeros((shape, shape, shape), dtype=torch.long)
+            return image, label
+
+    def Scanloader(self, db_file, label_type, num_cubes):
+        # Return a dummy Scanloader instance with a get_loaders method.
+        class DummyScanloader:
+            def __init__(self):
+                self.device = torch.device("cpu")
+            def get_loaders(self, batch_size=1, shuffle=True, num_workers=0):
+                dataset = DummyLoader.DummyDataset()
+                loader = torch.utils.data.DataLoader(dataset, batch_size=batch_size, shuffle=shuffle, num_workers=num_workers)
+                # For simplicity, return the same loader for train, valid, and test.
+                return loader, loader, loader
+        return DummyScanloader()
+
+# Dummy Dice: Returns a constant tensor.
+class DummyDice:
+    @staticmethod
+    def faster_dice(x, y, labels, fudge_factor=1e-8):
+        # Return a dummy tensor (for example, a single value).
+        return torch.tensor([1.0])
+
+# Dummy Logger: Does nothing (or prints) for testing.
+class DummyLogger:
+    def log_message(self, message, level=logging.INFO):
+        # For testing, we can simply pass or print(message)
+        pass
+
+# ----------------------------
+# Fixture for training class setup
+# ----------------------------
+@pytest.fixture
+def dummy_training_setup(tmp_path):
+    """
+    Create a temporary environment with a dummy modelAE.json file,
+    and instantiate the training class with dummy dependencies.
+    """
+    # Create a temporary directory for the training "path"
+    temp_dir = tmp_path / "dummy_dir"
+    temp_dir.mkdir()
+
+    # Create a dummy modelAE.json file (content is irrelevant for the dummy model)
+    modelAE_path = temp_dir / "modelAE.json"
+    with open(modelAE_path, "w") as f:
+        f.write("{}")  # empty JSON since DummyMeshNet doesn't use it
+
+    # Create a dummy output directory for saving the model
+    output_dir = temp_dir / "output"
+    output_dir.mkdir()
+
[…]
