+++ tests/conftest.py
@@ -1,27 +1,52 @@
 import sys
 from types import ModuleType
 
-# Check if nvflare is already installed; if not, create dummy modules.
+# If nvflare is not installed, create dummy modules.
 if "nvflare" not in sys.modules:
-    # Create the nvflare module.
+    # Create the top-level nvflare module.
     nvflare = ModuleType("nvflare")
-    apis = ModuleType("nvflare.apis")
-    fl_constant = ModuleType("nvflare.apis.fl_constant")
-    fl_context = ModuleType("nvflare.apis.fl_context")
+    sys.modules["nvflare"] = nvflare
 
-    # Create a dummy FLContextKey with the attribute CLIENT_NAME.
+    # Create the nvflare.apis module.
+    apis = ModuleType("nvflare.apis")
+    sys.modules["nvflare.apis"] = apis
+
+    # Create dummy module for nvflare.apis.fl_constant.
+    fl_constant = ModuleType("nvflare.apis.fl_constant")
+    # Define a dummy FLContextKey with a CLIENT_NAME attribute.
     class DummyFLContextKey:
         CLIENT_NAME = "CLIENT_NAME"
     fl_constant.FLContextKey = DummyFLContextKey
+    sys.modules["nvflare.apis.fl_constant"] = fl_constant
 
-    # Create a dummy FLContext class (if needed by your tests).
+    # Create dummy module for nvflare.apis.fl_context.
+    fl_context = ModuleType("nvflare.apis.fl_context")
     class DummyFLContext:
         pass
     fl_context.FLContext = DummyFLContext
-
-    # Register these dummy modules in sys.modules.
-    sys.modules["nvflare"] = nvflare
-    sys.modules["nvflare.apis"] = apis
-    sys.modules["nvflare.apis.fl_constant"] = fl_constant
     sys.modules["nvflare.apis.fl_context"] = fl_context
+
+    # Create dummy module for nvflare.apis.executor.
+    fl_executor = ModuleType("nvflare.apis.executor")
+    # Define a dummy Executor base class.
+    class DummyExecutor:
+        def __init__(self):
+            pass
+    fl_executor.Executor = DummyExecutor
+    sys.modules["nvflare.apis.executor"] = fl_executor
+
+    # Create dummy module for nvflare.apis.shareable.
+    fl_shareable = ModuleType("nvflare.apis.shareable")
+    # Define a dummy Shareable class (here we simply subclass dict).
+    class DummyShareable(dict):
+        pass
+    fl_shareable.Shareable = DummyShareable
+    sys.modules["nvflare.apis.shareable"] = fl_shareable
+
+    # Optionally, create dummy module for nvflare.apis.signal if needed.
+    fl_signal = ModuleType("nvflare.apis.signal")
+    class DummySignal:
+        pass
+    fl_signal.Signal = DummySignal
+    sys.modules["nvflare.apis.signal"] = fl_signal
 
