+++ requirements.txt
@@ -1,3 +1,4 @@
 pytest==8.3.4
 torch==2.0.1
 numpy==1.22.0
+nibabel
