### 2025-05-08 #7714 @likeajumprope
:waive @mikemcka based on your expertise, would you be willing to take this paper on for review?
