### 2025-11-05 #9357 @editorialbot
[REVIEW]: MeshFL: A Decentralized MeshNet Framework for 3D Brain MRI Segmentation
**Submitting author:** @Mmasoud1 (<a href="http://orcid.org/0000-0002-5365-242X">Mohamed Masoud</a>)
**Repository:** https://github.com/Mmasoud1/MeshFL
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
