### 2026-01-24 #9357 @paquiteau
@crvernon I have had similar issues for the installation as reported by @donishadsmith  in https://github.com/Mmasoud1/MeshFL/issues/2. I would also advise on properly packaging the repos, following guidelines from https://packaging.python.org/en/latest/, this also limits the possibilities to properly assess the functionality of the software.

Regarding the rest of the checklist, I am unfortunately not confident checking the "Substantial Scholarly Effort":

After exploring the codebase, the original contribution is quite minimal (most of material committed to the repo is only to make NVFlare works), for instance, the app/code folder only contains 780 lines of Python codes (according to `cloc`), including the definition of MeshNet and backup*s* of `meshnet_executor.py` (they should be removed btw, SVC software exists for a reason). Similarly the commit history is quite short once we don't consider the ones related to the JOSS paper (only ~30 commits until v1.0.0).

Overall, this contribution is lacking an original contribution in some form, be it good documentation and examples educating the user on the use of this software or original application of the NVFlare framework and MeshNet.
