### 2026-01-25 #9357 @Mmasoud1
@crvernon @paquiteau  thank you for feedback,  I appreciate the  opportunity to clarify  the  scope and intent of the contribution:

To clarify the scope, the  780 LOC  figure appears to be based on a narrow subfolder view. Even excluding **meshnet_executor.py**, the core modules/configs that implement the federated workflow (aggregator, data loading, path resolution, logging, model/config, and Dice utilities) already total  around 612 lines. Around 8 helper files (e.g. loader.py, dist.py  etc ) had been coded  besides meshnet_executor.py,   which together contribute to the overall federated training pipeline.  I will clarify this in the documentation to avoid any possible confusion, so thank you to bringing this to my attention.

I agree that commit count are imperfect indicators of scholarly contribution especially for systems integration work. But that is beacsue many lines of the  workflow was early developed  and tested iteratively in a local /cluster environment and then pushed to GitHub in consolidated updates, which makes the public commit history appear shorter than the actual development cycle. Going forward, I will improve transparency with clearer release notes.

The **Substantial Scholarly Effort** in this work is not intended to be measured by model novelty or raw line count alone , but by the reproducible integration of a non trivial 3D medical-imaging model (i.e. MeshNet) into a federated learning workflow using NVFlare. The contribution focuses on enabling decentralized training,  gradient aggregation, evaluation,  for MRI segmentation under realistic privacy and governance constraints.  In real multi-institution settings (including pediatric imaging cohorts), data cannot be centralized due to privacy, governance, and jurisdictional constraints. A working FL example that supports local training, gradient aggregation, evaluation, and outputs enables sites to collaboratively improve segmentation models without sharing sensitive scans. This lowers the barrier for research groups across different hospitals and regions to adopt federated learning workflows for brain-region segmentation used in downstream developmental analyses.

I agree, however, that the current documentation/examples do not yet communicate this scope clearly. I will strengthen the documentation with an end-to-end usage guide (simulator run, expected outputs, and adaptation steps) and clean up redundant/back-up files so the contribution can be assessed more directly.

@paquiteau Please kindly  let me know if these clarification and  proposed changes would address your concern?
