+++ requirements.txt
@@ -1,4 +1,5 @@
-pytest==8.3.4
-torch==2.0.1
 numpy==1.22.0
+pandas==2.0.3
 nibabel
+torch==2.0.1
+nvflare==2.4.0
