### 2026-03-29 #9357 @Mmasoud1
Thank you @crvernon  for the following up, and thanks for @paquiteau and @donishadsmith for feedback and patience.

I apologize for the delay in my response due to international travels over the past month. I have been actively working on addressing the comments over the past couple of days and addressed most comments.

Here is a quick update on the current status:

* I am restructuring the installation process to make it more straightforward, including cleaning up and reorganizing the `requirements.txt` (separating core vs testing dependencies as suggested by @donishadsmith).
* Missing dependencies such as `nibabel` are being added and verified.
* Python version compatibility is being clarified and documented in Wiki Setup section
* Addressed the notes on Git LFS requirements for the example datasets in Wiki Setup section.
* I am also preparing a Google Colab demo to make the project easier to run and reproduce, the colab almost done and will be added shortly.

Due to time constraints today, I plan to:

* Push the remain fixes within couple of days.
* Add the Colab demo and finalize documentation shortly after.

My goal is to have everything addressed and cleaned up this week with full report.  Thank you again for your valuable feedback and patience.
