### 2026-03-31 930c75b
Fix param.grad handling
