+++ app/code/executor/meshnet_executor.py
@@ -52,16 +52,6 @@ class MeshNetExecutor(Executor):
         self.model.to(self.device)
 
 
-        # # Check if the class variable initial_weights is None  ( Failded becasue both sites give copy)
-        # if MeshNetExecutor.initial_weights is None:
-        #     # If this is the first site, store the model's initial weights
-        #     MeshNetExecutor.initial_weights = {key: value.clone() for key, value in self.model.state_dict().items()}
-        #     self.logger.log_message(f"Initial weights copied from this site")
-        # else:
-        #     # If weights have already been initialized, load them into this site's model
-        #     self.model.load_state_dict(MeshNetExecutor.initial_weights)
-        #     self.logger.log_message(f"Initial weights assigned to this site.")
-
 
         # Define the file path for the weights   
         initial_weights_file_path = os.path.join(os.path.dirname(__file__), "initial_weights.pth") 
@@ -141,18 +131,6 @@ class MeshNetExecutor(Executor):
 
 
 
-
-    # def set_seed(self, seed):
-    #     # Set the random seed for reproducibility
-    #     random.seed(seed)
-    #     np.random.seed(seed)
-    #     torch.manual_seed(seed)
-    #     if torch.cuda.is_available():
-    #         torch.cuda.manual_seed(seed)
-    #         torch.cuda.manual_seed_all(seed)
-    #     torch.backends.cudnn.deterministic = True  # Ensure deterministic behavior
-    #     torch.backends.cudnn.benchmark = False
-
     def execute(
         self,
         task_name: str,
@@ -163,21 +141,7 @@ class MeshNetExecutor(Executor):
         
         # Get site name
         self.site_name = fl_ctx.get_prop(FLContextKey.CLIENT_NAME)
-
-
-        # Check for initial model broadcast 
-        # (((( FOR FUTUR USE))))  as practicall solution to share init weights between sites at begining. 
-        # if task_name == "initialize_model":
-        #     self.logger.log_message(f"{self.site_name}-initialize_model weights called")
-        #     model_weights = shareable["model_weights"]
-        #     self.load_model_weights(model_weights)
-        #     return Shareable()
-
-        # if "initial_weights" in shareable:
-        #     self.logger.log_message(f"{self.site_name} - Received initial weights from the server.")
-        #     initial_weights = shareable["initial_weights"]
-        #     self.model.load_state_dict(initial_weights)  # Load the initial weights
-        #     return Shareable()  # Acknowledge receipt        
+  
 
 
         # Initialize data loader and other site-specific configurations once
@@ -232,160 +196,7 @@ class MeshNetExecutor(Executor):
             self.apply_gradients(aggregated_gradients, fl_ctx)
             return Shareable()
 
-
-
-    # def train_and_get_gradients_old(self, fl_ctx):
-    #     # Set the model to training mode, so dropout is activated
-    #     # and BatchNorm normalizes the input based on the statistics of the current mini-batch.
-    #     self.model.train()
-
-    #     # Initialize accumulators for the loss and gradients
-    #     total_loss = 0.0
-
-    
-
-    #     # Training loop for one epoch (full pass through the dataset)
-
-    #     for batch_id, (image, label) in enumerate(self.trainloader):
-
-    #         # Moving data to the correct device (CPU or GPU)
-    #         image, label = image.to(self.device), label.to(self.device)
-
-    #         self.optimizer.zero_grad() #  It resets/clears the gradients of all model parameters (i.e., weights).
-    #         # PyTorch by defualt adds new gradients to any existing gradients, to prevent this accumulation of gradients 
-    #         # from previous iterations, manually set them to zero at the beginning of each new training iteration.
-
-
-    #         # Mixed precision and checkpointing
[…]
