+++ app/code/aggregator/get_global_average.py
@@ -1,7 +1,0 @@
-from typing import List, Dict
-
-
-def get_global_aggregated_gradients(aggregated_gradients):
-    # May remain for additional post-processing of gradients, if necessary, 
-    # but the core logic is now in gradient_aggregator.py.
-    return aggregated_gradients
