### 2026-04-03 #9357 @Mmasoud1
Thank you @crvernon for the follow-up, and many thanks to @paquiteau and @donishadsmith for your constructive feedback.

To the best I have addressed the respected  reviewers comments with substantial improvements to installation, documentation, usability, and clarity of the contribution based on their valuable comments.  Below is a concise summary of the final updates:

## 1- Installation and Dependencies:

- Added a clean [`requirements.txt`](https://github.com/Mmasoud1/MeshFL/blob/main/requirements.txt) at the repository root for simplified installation:
```
pip install -r requirements.txt
```
- Added missing dependencies such as `nibabel`

- Separated testing dependencies into [`requirements-dev.txt`](https://github.com/Mmasoud1/MeshFL/blob/main/requirements-dev.txt)

- Fixed invalid Pytorch distribution in requirements and show in the demo a working dependency versions and installation example.

- Modified [Setup page](https://github.com/Mmasoud1/MeshFL/wiki/Setup) and documented Python compatibility (recommended Python 3.8–3.10 due to NVFlare constraints). Added instructions for:
  - Git LFS (required for dataset download)
  - Installation (Linux + guidance for Windows)

## 2- Documentation and Demo

To address concerns regarding clarity and usability, I significantly expanded documentation and added a full demo:

###   Demo Notebook + Explanation
A complete end-to-end [Demo](https://github.com/Mmasoud1/MeshFL/wiki/MeshFL-Demo) has been added showing:
- decentralized training behavior
- convergence (Loss & Dice)
- learning rate scheduling
- segmentation inference and overlay visualization

Please check :
[Demo Notebook](Examples/MeshFL_Demo.ipynb)

[Run on Colab](https://colab.research.google.com/drive/18xjmBMR7EhuFVyhqhHuoGoCLJUuCwUA2?usp=sharing)

###  Training Behavior and Concepts
Clarified:
- decentralized learning dynamics
- training configuration and rationale

Please check [Training concepts page](https://github.com/Mmasoud1/MeshFL/wiki/Training-Configuration-&-Concepts)

### Simulation Outputs
Added detailed explanation of:
- logs
- expected outputs
- how to interpret training results

Please check [Simulation output page](https://github.com/Mmasoud1/MeshFL/wiki/Simulation-Output-and-Logs)

### CI and Testing
Updated CI documentation and dependency structure after separated requirements and requirements-dev, please check [CI page](
https://github.com/Mmasoud1/MeshFL/wiki/Continuous-Integration-(CI))

Real example of automated testing can be found with this last pushing action [here](https://github.com/Mmasoud1/MeshFL/actions/runs/23926527031/job/69784442782) with subsection "**Run tests with pytest**"

Also testing can run locally following instructions from [Unit Testing page](https://github.com/Mmasoud1/MeshFL/wiki/Unit-Testing)

## 3- Repository Cleanup

- Removed redundant and backup files
- Improved repository structure

## 4- Clarification on Contribution

To better communicate the scholarly contribution:

- The work focuses on **integrating MeshNet into a decentralized NVFlare workflow**
- Provides a **reproducible pipeline for privacy-preserving 3D MRI segmentation**
- Demonstrates **multi-site training, aggregation, and inference**
- Includes a **complete working example**, addressing the reviewer’s concern about usability and educational value

## 5- Dataset Documentation and Handling Improvements

- Added a dedicated **[Dataset Wiki page](https://github.com/Mmasoud1/MeshFL/wiki/Data)** describing:
  - dataset origin (Mindboggle subset)
  - database schema and others

- Clarified that MeshFL uses a **SQLite-based dataset with BLOB storage**, where MRI volumes and labels are stored directly inside the database rather than as file paths.

- Added a **dataset read notebook and colab**:
  - [MeshFL_dataset_read.ipynb](https://github.com/Mmasoud1/MeshFL/blob/main/Examples/dataset_read/MeshFL_dataset_read.ipynb)
  - [Colab](https://colab.research.google.com/drive/181xz8J3vZHupZC8og2ZYcauxfLZYYiNR?usp=sharing) version
  - Includes visualization of MRI, labels, and overlays directly from the database

- Added a **dataset creation notebook and colab**:
  - [MeshFL_dataset_create.ipynb](https://github.com/Mmasoud1/MeshFL/blob/main/Examples/dataset_create/MeshFL_dataset_create.ipynb)
  - [Colab](https://colab.research.google.com/drive/1a919qB1G9jfoIeEBOqUhmimmeZuqyJ_M?usp=sharing) version
  - Demonstrates how to convert raw NIfTI files into a MeshFL-compatible SQLite database

- Clarified **multi-site dataset organization**, where:
  - each site uses an independent local database
  - no raw data is shared
  - only model updates are exchanged

These additions improve transparency of the data pipeline and make it easier for users to understand, reproduce, and extend MeshFL with their own datasets.

## 6- MeshFL - v2.0.0 Release Status

The repository has now been cleaned, documented, and  validated with demo and installation improvements. It is ready for **v2.0.0 release** upon respected reviewers approval.

I truly appreciate @paquiteau and @donishadsmith feedback, it significantly improved the clarity and usability of this work.
